FIx dialog null owner bug. Closes #3.
Switched to using ComboBox since the JComboBox had a deprecated warning.
That required a fixed input size, made it 200 pixels, arbitrarily.
The bug was that were not setting owner if nothing was ever selected.
Instead of owner I made the orgList ComboBox a class property and am
querying that now in the do ok action.