Fix auto-login with .floorc.json is not present.
The last commit couldn't fix the one case where common code needed to open a browser, which was in the "I already have an account" setup flow.  To make the IntelliJ-specific browser opening code available to core libraries without actually referencing IntelliJ's library from common code, I've created a replaceable "BrowserOpener" singleton that uses the default java.awt.Desktop behavior most of the time, but gets replaced by an "IntelliBrowserOpener" when this plugin is initialized which will try the IntelliJ open-in-browser function first.